floating point32 performance test

+++ pattern.cpp
@@ -32,9 +32,9 @@ int pattern::load(const char * const pFilename)
 
     this->width = header.width;
     this->height = header.height;
-    this->size = header.width * header.height * sizeof(double);
+    this->size = header.width * header.height * sizeof(float);
 
-    this->data = reinterpret_cast<double *>(malloc(this->size));
+    this->data = reinterpret_cast<float *>(malloc(this->size));
 
     if(!this->data){
         file.close();
@@ -70,14 +70,14 @@ int pattern::save(const char * const pFilename)
     return SUCCESS;
 }
 
-bool amp_sort(double i, double j)
+bool amp_sort(float i, float j)
 {
     return (fabs(i) < fabs(j));
 }
 
-double find_99_7_percentile(double * data, uint32_t count)
+float find_99_7_percentile(float * data, uint32_t count)
 {
-    std::vector<double> v(data, data + count);
+    std::vector<float> v(data, data + count);
     std::sort(v.begin(), v.end(), amp_sort);
     return fabs(v[254.0f/255.0f * count]);
 }
@@ -89,7 +89,7 @@ int pattern::export_bmp(const char * const pFilename)
 
     struct pixel32bpp * pixel_data;
 
-    double max_amp = abs(this->data[0]);
+    float max_amp = abs(this->data[0]);
 
     std::ofstream file(pFilename, std::ios::binary | std::ios::trunc);
     if(!file.is_open()) {
@@ -121,7 +121,7 @@ int pattern::export_bmp(const char * const pFilename)
 
     /* for(uint32_t c = 1; c < this->width * this->height; c++)
     {
-        double val = fabs(this->data[c]);
+        float val = fabs(this->data[c]);
         if(val > max_amp) {
             max_amp = val;
         }
@@ -137,7 +137,7 @@ int pattern::export_bmp(const char * const pFilename)
             uint32_t data_pos = x + y * this->width;
             uint32_t pixel_pos = x + (this->height - y - 1) * this->width;
 
-            double amp = fabs(this->data[data_pos]);
+            float amp = fabs(this->data[data_pos]);
 
             uint8_t val = amp > max_amp ? 0xff : amp / max_amp * 0xff;
             pixel_data[pixel_pos].raw = 0xffffffff;
@@ -160,9 +160,9 @@ int pattern::export_bmp(const char * const pFilename)
 }
 
 pattern::pattern(uint32_t const width, uint32_t const height)
-    :width(width), height(height), size(width * height * sizeof(double))
+    :width(width), height(height), size(width * height * sizeof(float))
 {
-    this->data = reinterpret_cast<double *>(malloc(this->size));
+    this->data = reinterpret_cast<float *>(malloc(this->size));
 }
 
 pattern::~pattern(void)
